bits: module resolution
