+++ bits/lib.rs
@@ -10,7 +10,6 @@ macro_rules! mask {
     )*)
 }
 
-mod bits;
 mod container;
 mod container_mut;
 mod count;
@@ -27,7 +26,6 @@ pub mod not;
 pub mod or;
 pub mod xor;
 
-pub use self::bits::Bits;
 pub use self::container::Container;
 pub use self::container_mut::ContainerMut;
 pub use self::count::Count;
@@ -36,3 +34,64 @@ pub use self::mask::Mask;
 pub use self::rank::Rank;
 pub use self::select::Select;
 pub use self::{lsb::Lsb, msb::Msb};
+
+pub trait Bits: Clone + Container + ContainerMut + Count + Rank + Excess + Select {
+    const BITS: usize;
+
+    #[doc(hidden)]
+    const SIZE: usize = Self::BITS / 8;
+
+    fn empty() -> Self;
+}
+
+macro_rules! ints_impl_bits {
+    ($( $Int:ty )*) => ($(
+        impl Bits for $Int {
+            const BITS: usize = <$Int>::BITS as usize;
+
+            #[inline]
+            fn empty() -> Self {
+                0
+            }
+        }
+
+    )*)
+}
+ints_impl_bits!(u8 u16 u32 u64 u128 usize);
+ints_impl_bits!(i8 i16 i32 i64 i128 isize);
+
+impl<B, const N: usize> Bits for [B; N]
+where
+    B: Copy + Bits,
+{
+    const BITS: usize = B::BITS * N;
+
+    #[inline]
+    fn empty() -> Self {
+        [B::empty(); N]
+    }
+}
+
+mod impl_bits {
+    use super::*;
+    use std::borrow::Cow;
+
+    impl<B: Bits> Bits for Box<B> {
+        const BITS: usize = B::BITS;
+        #[inline]
+        fn empty() -> Self {
+            Box::new(B::empty())
+        }
+    }
+
+    impl<'a, B> Bits for Cow<'a, B>
+    where
+        B: ?Sized + Bits,
+    {
+        const BITS: usize = B::BITS;
+        #[inline]
+        fn empty() -> Self {
+            Cow::Owned(B::empty())
+        }
+    }
+}
